downloader.urls/requirements.txt
imagehash>=4.3.2
Pillow>=11.0.0
requests>=2.32.0
numpy>=1.24.0
scipy>=1.10.0
PyWavelets>=1.4.0